Move all test utils to test package
COAND-852
