+++ example-app/src/androidTest/java/com/adyen/checkout/test/fake/FakeCheckoutConfigurationProvider.kt
@@ -6,13 +6,13 @@
  * Created by oscars on 6/6/2024.
  */
 
-package com.adyen.checkout.fake
+package com.adyen.checkout.test.fake
 
 import com.adyen.checkout.components.core.Amount
 import com.adyen.checkout.components.core.CheckoutConfiguration
 import com.adyen.checkout.core.Environment
 import com.adyen.checkout.example.ui.configuration.ConfigurationProvider
-import com.adyen.checkout.server.CheckoutMockWebServer
+import com.adyen.checkout.test.server.CheckoutMockWebServer
 import java.net.URL
 import java.util.Locale
 import javax.inject.Inject
