+++ example-app/src/androidTest/java/com/adyen/checkout/test/server/service/MockBackendService.kt
@@ -3,12 +3,12 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by oscars on 5/6/2024.
+ * Created by oscars on 6/6/2024.
  */
 
-package com.adyen.checkout.server.service
+package com.adyen.checkout.test.server.service
 
-import com.adyen.checkout.util.JsonFileReader
+import com.adyen.checkout.test.util.JsonFileReader
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.RecordedRequest
 import java.net.HttpURLConnection.HTTP_OK
