+++ example-app/src/androidTest/java/com/adyen/checkout/test/server/service/MockCheckoutService.kt
@@ -3,10 +3,10 @@
  *
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  *
- * Created by oscars on 5/6/2024.
+ * Created by oscars on 6/6/2024.
  */
 
-package com.adyen.checkout.server.service
+package com.adyen.checkout.test.server.service
 
 import okhttp3.mockwebserver.MockResponse
 import okhttp3.mockwebserver.RecordedRequest
